third_party/blink/web_tests/fast/serviceworker/serviceworkercontainer-interface.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_true('ServiceWorkerContainer' in window,
              'the constructor should be defined');
  assert_throws(new TypeError, function() { new ServiceWorkerContainer(); },
                'the constructor should not be callable with "new"');
  assert_throws(new TypeError, function() { ServiceWorkerContainer(); },
                'the constructor should not be callable');
}, 'ServiceWorkerContainer interface');
</script>
